issue/46: Move BINARY_DESCRIPTOR to binary.h, remove try-catch in BinaryInfo create(), remove redundant binary.h under binary/cpu/, etc.
